bench_nexus functionality returned back to module
